nbd: fix type conversion error when doing ioctl
This is reproducible on e6500 platforms, with 64bit kernel and 32bit rootfs
(tested on T2080RDB). Not reproducible on 64bit kernel with 64bit rootfs
(testeed on T4240RDB-64b).

(backported from commit 3f4726596dafd2e27485e51f4cc4a2363f48d4a3)
Reviewed-on: http://git.am.freescale.net:8181/34258
Tested-by: Mihai Caraman <mihai.caraman@freescale.com>

+++ nbd.c
@@ -640,9 +640,9 @@ int nbd_init(int fd, int csock, uint32_t flags, off_t size, size_t blocksize)
         return -serrno;
     }
 
-        TRACE("Setting size to %zd block(s)", (size_t)(size / blocksize));
+    TRACE("Setting size to %zd block(s)", (size_t)(size / blocksize));
 
-    if (ioctl(fd, NBD_SET_SIZE_BLOCKS, size / blocksize) < 0) {
+    if (ioctl(fd, NBD_SET_SIZE_BLOCKS, (size_t)(size / blocksize)) < 0) {
         int serrno = errno;
         LOG("Failed setting size (in blocks)");
         return -serrno;
